# Optimize _is_python_file check